backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-multipart==0.0.6
minio==7.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
transformers==4.37.2
torch==2.1.2
sentencepiece==0.1.99
pytesseract==0.3.10
pdf2image==1.17.0
numpy==1.26.4
spacy==3.7.4
# After install: python -m spacy download it_core_news_lg
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.4.10
python-docx==1.2.0
openpyxl==3.1.5
openai>=1.30.0
python-dotenv>=1.0.0
